+++ tomcat/src/main/java/org/apache/catalina/Manager.java
@@ -1,10 +1,8 @@
 package org.apache.catalina;
 
-import jakarta.servlet.http.HttpSession;
-
 import java.io.IOException;
 
-import org.apache.coyote.http11.Session;
+import org.apache.catalina.session.Session;
 
 /**
  * A <b>Manager</b> manages the pool of Sessions that are associated with a
