+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -17,8 +17,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
+import org.apache.catalina.session.Session;
+import org.apache.catalina.session.SessionManager;
 import org.apache.coyote.Processor;
-import org.apache.coyote.http11.domain.Cookie;
+import org.apache.coyote.http11.cookie.Cookie;
 import org.apache.coyote.http11.io.HttpResponseWriter;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.request.HttpRequestBody;
