+++ tomcat/src/main/java/org/apache/coyote/http11/cookie/Cookie.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.domain;
+package org.apache.coyote.http11.cookie;
 
 import java.util.HashMap;
 import java.util.Map;
